chore: rename types to type
