+++ package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "@nuxtjs/i18n": "^8.0.0-rc.4",
+    "@types/validator": "^13.11.2",
     "nuxt": "^3.7.4",
     "prettier": "^2.8.8",
     "sass": "^1.68.0",
