+++ src/ts/type/AppConfig.ts
@@ -1,0 +1,9 @@
+type AppConfig = {
+  publicUrl: string
+  appVersion: string
+  apiBaseUrl: string
+  cookieAuhtName: string
+  appName: string
+}
+
+export default AppConfig
