+++ src/ts/type/Colors.ts
@@ -1,0 +1,23 @@
+export type BaseColors = {
+  background: string
+  surface: string
+  primary: string
+  secondary: string
+  success: string
+  warning: string
+  error: string
+  info: string
+}
+
+export type OnColors = {
+  'on-background': string
+  'on-surface': string
+  'on-primary': string
+  'on-secondary': string
+  'on-success': string
+  'on-warning': string
+  'on-error': string
+  'on-info': string
+}
+
+export type themesColors = Readonly<Partial<BaseColors> & Partial<OnColors>>
