+++ src/ts/type/Jwt.ts
@@ -1,0 +1,10 @@
+import User from './User'
+
+type JwtType = {
+  exp: number
+  data: User
+  iat: number
+  [key: string]: unknown
+}
+
+export default JwtType
