+++ src/ts/type/User.ts
@@ -1,0 +1,3 @@
+export default interface User {
+  username: string
+}
