--- src/ts/types/AppConfig.ts
@@ -1,9 +1,0 @@
-type AppConfig = {
-  publicUrl: string
-  appVersion: string
-  apiBaseUrl: string
-  cookieAuhtName: string
-  appName: string
-}
-
-export default AppConfig
