--- src/ts/types/Jwt.ts
@@ -1,10 +1,0 @@
-import User from './User'
-
-type JwtType = {
-  exp: number
-  data: User
-  iat: number
-  [key: string]: unknown
-}
-
-export default JwtType
