--- src/ts/types/User.ts
@@ -1,3 +1,0 @@
-export default interface User {
-  username: string
-}
